+++ gradle-jdks-setup/src/test/java/com/palantir/gradle/jdks/setup/CertificateUtilsTest.java
@@ -22,29 +22,30 @@ import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import org.junit.jupiter.api.Test;
 
-public class CaResourcesTest {
+public class CertificateUtilsTest {
 
     @Test
     void handles_just_whitespace_truststore() {
-        assertThat(CaResources.parseCerts("         \n      ".getBytes(StandardCharsets.UTF_8)))
+        assertThat(CertificateUtils.parseCerts("         \n      ".getBytes(StandardCharsets.UTF_8)))
                 .isEmpty();
     }
 
     @Test
     void handles_whitespace_and_comments_between_certs() throws IOException {
-        assertThat(CaResources.parseCerts(certsFromResources("amazon-cas-with-whitespace-between.pem")))
+        assertThat(CertificateUtils.parseCerts(certsFromResources("amazon-cas-with-whitespace-between.pem")))
                 .hasSize(3);
     }
 
     @Test
     void does_not_explode_when_given_certs_with_duplicate_extensions() throws IOException {
-        assertThat(CaResources.parseCerts(certsFromResources("apple-kdc-cert-with-duplicate-extension.pem")))
+        assertThat(CertificateUtils.parseCerts(certsFromResources("apple-kdc-cert-with-duplicate-extension.pem")))
                 .isEmpty();
     }
 
     @Test
     void does_not_explode_when_given_certs_with_incorrect_vesion() throws IOException {
-        assertThat(CaResources.parseCerts(certsFromResources("strongloop-cert-with-v3-extensions-but-v1-version.pem")))
+        assertThat(CertificateUtils.parseCerts(
+                        certsFromResources("strongloop-cert-with-v3-extensions-but-v1-version.pem")))
                 .isEmpty();
     }
